#1700: move writeRPMConfig and firewallFilterTermExpansionCount helpers
Relocate the two remaining package-level free functions out of
server_show.go to sit with their consumers: writeRPMConfig joins
showRPM in server_show_security_text.go; firewallFilterTermExpansionCount
joins the firewall-filter renderers in server_show_firewall.go. Pure
code-motion, no behavior change.

server_show.go is now 479 LOC (was 2006) — the dispatcher facade plus
the default/log branches only. Full pkg/grpcapi suite green; ShowText
golden guard byte-identical.
